[tuya] Refactor C++ code into Tuya base class and TuyaUART subclass
To allow for other controller implementations using the same Tuya base
class and exchanging datapoints using the existing sensors and other
components of the 'tuya' platform, separate out the base class and DP
handling from the code which is specific to this particular UART-based
controller.

The existing user configs don't change; only the C++. The config
generator emits a 'TuyaUART' now instead of the base 'Tuya' class.

The sensors aren't changed because they only ever used the generic
datapoint handling code which is now in the base Tuya class anyway,
which is kind of the point.

The one real change that isn't just moving code from one file to another
is the addition of a handle_datapoint_() method to the base class, which
is now invoked from the handle_datapoints() method of TuyaUART class,
having pulled that generic code out of the UART implementation.