[tuya] Refactor config generator to allow for extensibility
A previous commit reworked the C++ code to allow for subclasses of the
Tuya base class which communicate over different methods. This extends
the config generation to support it too.

The key change is to handle autoloading. The other controller component
needs to autoload 'tuya' to get the *base* class, but doesn't want to
get a TuyaUART instantiated along with it. So the tuya code generator
can detect if the config is an instance of the AutoLoad class, and
simply refrain from doing so.

The config validator needs to do the same check, because otherwise the
uart.UART_DEVICE_SCHEMA will complain if there isn't an actual UART in
the system.

Finally, split out the config schema and code generation for the options
which reside in the base class, to helpers so that the other
implementations can use those.